Разделил исполняемый файл на "интерфейс" и библиотеку. Код так и не тестировал. 100% не рабочий
